% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fixef_table.R
\name{format_fixef_table}
\alias{format_fixef_table}
\title{Title}
\usage{
format_fixef_table(model, type = "cond")
}
\arguments{
\item{model}{input model}

\item{type}{type of parameters to export}
}
\value{
a tibble
}
\description{
Title
}
