% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_durations}
\alias{region_durations}
\title{Brushing durations for 13 subjects, 10 sessions each, by region}
\format{
TODO: complete this
}
\usage{
region_durations
}
\description{
Brushing durations for 13 subjects, 10 sessions each, by region
}
\keyword{datasets}
