% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ranef_table.R
\name{format_ranef_table}
\alias{format_ranef_table}
\title{Title}
\usage{
format_ranef_table(model, comp = "cond", combine_stats = TRUE)
}
\arguments{
\item{combine_stats}{}
}
\value{

}
\description{
Title
}
